pyproject.toml
[project]
name = "currency-exchange-api"
version = "0.1.0"
description = "Currency Exchange API is a Django-based application providing currency exchange rates, user balance management, and transaction history. The API integrates with an external ExchangeRate API, uses JWT authentication, and is containerized using Docker and managed via Poetry."
authors = [
    {name = "Andrey Dementyev",email = "dementyev.andreyy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"

[tool.ruff]
line-length = 91

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]
ignore = [
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "D",  # pydocstyle
    "F401",
    "B904",
    "N802",
    "B008",
    "F821",
    "I001",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
